+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * **trader:** added upgrade path for saved data.
 * **trader:** added buffer for minimum trade cost.
 * **trader:** compare open trades to the source strategy.
+* **trader:** added notifier level configuration.
 
 ## [1.1.0] (2021-08-09)